+++ CHANGELOG.rst
@@ -54,6 +54,8 @@ odd bug fix may not get mentioned here at all.
   - Feature: changed the location where missing attributes are reported. They
     now appear in the start label of an element.
 
+  - Feature: support for arrow up and arrow down to move the caret.
+
   - GUI Fix: When the user would use the down arrow to navigate the options of a
     completion menu, the focus would be lost from the document and would not be
     regained when the user closes the completion menu. This made further typing